refactoring name HTMLParser to HtmlParser
