+++ web-crawler-core/src/main/java/pl/edu/agh/student/oop/webcrawler/core/parser/HtmlParser.java
@@ -9,11 +9,11 @@ import org.jsoup.select.NodeVisitor;
 import java.util.ArrayList;
 import java.util.List;
 
-public class HTMLParser {
+public class HtmlParser {
 
     private Document doc;
 
-    public HTMLParser(Document doc) {
+    public HtmlParser(Document doc) {
         this.doc = doc;
     }
